// kadmium-dmx.reactui/src/generated/AddUniverse.ts
/* tslint:disable */
/* eslint-disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: AddUniverse
// ====================================================

export interface AddUniverse_addUniverse {
  __typename: "activeUniverse";
  name: string;
}

export interface AddUniverse {
  addUniverse: AddUniverse_addUniverse;
}

export interface AddUniverseVariables {
  name: string;
  universeId: number;
}