// kadmium-dmx.reactui/src/generated/DeleteUniverse.ts
/* tslint:disable */
/* eslint-disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: DeleteUniverse
// ====================================================

export interface DeleteUniverse {
  deleteUniverse: number;
}

export interface DeleteUniverseVariables {
  universeId: number;
}